use browserify for bundling npm modules to a single js file
