+++ .gitignore
@@ -3,3 +3,4 @@ bower_components
 node_modules
 npm-debug.log
 gh-pages
+dist
