+++ gulpfile.js
@@ -1,8 +1,8 @@
 var gulp = require('gulp');
 
-var buildDir = './gh-pages/';
+var buildDir = './dist/';
 
-gulp.task('build', function() {
+gulp.task('build:copy', function() {
   gulp.src('src/index.*').pipe(gulp.dest(buildDir));
   gulp.src('src/nls/locale-*.json').pipe(gulp.dest(buildDir + '/nls/'));
 });
