+++ package.json
@@ -5,8 +5,9 @@
   "main": "index.js",
   "scripts": {
     "test": "exit 1",
-    "lint": "jshint src/*.js && jscs src/*.js",
-    "build": "gulp build && gulp build:vendor"
+    "lint": "jshint src/js && jscs src/js",
+    "build": "gulp build:copy && gulp build:vendor && npm run build:bundle",
+    "build:bundle": "browserify src/js/index.js -d | exorcist dist/index.js.map > dist/index.js"
   },
   "repository": {
     "type": "git",
@@ -19,7 +20,10 @@
   },
   "homepage": "https://github.com/freifunk-berlin/firmware-wizard-frontend",
   "devDependencies": {
-    "gulp": "^3.9.0"
+    "browserify": "^11.2.0",
+    "exorcist": "^0.4.0",
+    "gulp": "^3.9.0",
+    "ip": "^1.0.1"
   },
   "dependencies": {
     "jscs": "^2.2.0",
