+++ src/js/index.js
@@ -4,6 +4,8 @@ var wizard = angular.module('WizardApp', [
   'ui.bootstrap', 'ngAnimate', 'leaflet-directive', 'pascalprecht.translate'
 ]);
 
+require('./services')(wizard);
+
 wizard.controller('WizardCtrl', [
   '$scope', 'leafletData', '$http', '$filter', 'downloadFile', '$translate',
   'jsonrpc',
@@ -350,47 +352,6 @@ wizard.config(function($translateProvider) {
     }
   );
 });
-
-wizard.factory('jsonrpc', ['$http', '$q', function($http, $q) {
-  var jsonrpc = {};
-  var call = function(session, object, method, args) {
-    var deferred = $q.defer();
-
-    $http.post(jsonrpc.apiUrl, {
-      jsonrpc: '2.0',
-      id: 1,
-      method: 'call',
-      params: [session, object, method, args]
-    })
-      .success(function(data) {
-        if (data.error) {
-          return deferred.reject('JSON RPC Error: ' + data.error.message +
-                          ' (code ' + data.error.code + ')');
-        }
-        deferred.resolve(data.result[1]);
-      })
-      .error(function(data) {
-        deferred.reject(data);
-      });
-
-    return deferred.promise;
-  };
-
-  jsonrpc.login = function(apiUrl, username, password) {
-    jsonrpc.apiUrl = apiUrl;
-    return call('00000000000000000000000000000000', 'session', 'login',
-         {'username': 'root', 'password': 'doener', 'timeout': 3600})
-      .then(function(data) {
-        jsonrpc.session = data.ubus_rpc_session;
-      });
-  };
-
-  jsonrpc.call = function(object, method, args) {
-    return call(jsonrpc.session, object, method, args);
-  };
-
-  return jsonrpc;
-}]);
 
 wizard.factory('downloadFile',
   ['$document', 'base64encodeFilter', function($document, base64encode) {
