+++ src/js/services/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = function(module) {
+  require('./jsonrpc')(module);
+};
